Migrated the pretty printing code from VerificationReport.scala to SelfPrettyPrinter.scala
